+++ examples/blog/trilha_gen.go
@@ -24,6 +24,7 @@ import (
 	app_docs_path__ "github.com/emersonjoe/trilha/examples/blog/app/docs/path__"
 	app_documentos "github.com/emersonjoe/trilha/examples/blog/app/documentos"
 	app_documentos_planilha "github.com/emersonjoe/trilha/examples/blog/app/documentos/planilha"
+	app_documentos_resumo "github.com/emersonjoe/trilha/examples/blog/app/documentos/resumo"
 	app_legado_ "github.com/emersonjoe/trilha/examples/blog/app/legado-"
 	app_legado__legado "github.com/emersonjoe/trilha/examples/blog/app/legado-/legado"
 	app_legado__legado_apagar "github.com/emersonjoe/trilha/examples/blog/app/legado-/legado/apagar"
@@ -169,6 +170,12 @@ func newApp() *trilha.App {
 			"GET":  app_documentos_planilha.GET,
 			"POST": app_documentos_planilha.POST,
 		},
+		Middlewares: []trilha.MiddlewareFunc{app.Middleware},
+	})
+	a.Register(trilha.Route{
+		Pattern:     "/documentos/resumo",
+		Page:        app_documentos_resumo.Page,
+		Layouts:     []trilha.LayoutFunc{app.Layout},
 		Middlewares: []trilha.MiddlewareFunc{app.Middleware},
 	})
 	a.Register(trilha.Route{